Move ecdsa error code to submodule
Move the error code out of the `ecdsa/mod.rs` file and into a private
`error` submodule. Add public re-export so this not a breaking change.

Code move only plus import/export changes to build cleanly. No other
changes.
